Auto: 2026-01-14 17:21

+++ jpa-crud/src/main/java/com/example/myapp/controller/UserController.java
@@ -23,9 +23,9 @@ public class UserController {
 
     // ==== 2. 회원 단건 조회 ====
     @GetMapping("/{id}")    // GET /api/users/{id}
-    public User get(@PathVariable int id){
+    public User get(@PathVariable long id){
         return userService.get(id);
-        // 
+        //
     }
